css-align-3_dev/html/place-content-shorthand-003.htm

<!DOCTYPE html>
<html><head><title>CSS Box Alignment: place-content shorthand - initial value</title>
<link href="mailto:jfernandez@igalia.com" rel="author" title="Javier Fernandez">
<link href="http://www.w3.org/TR/css3-align/#propdef-place-content" rel="help">
<meta content="Check that place-content's 'initial' value expands to 'align-content' and 'justify-content'." name="assert">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
</head><body><div id="log"></div>
<script>
    var div = document.createElement("div");
    document.body.appendChild(div);
    div.style["align-content"] = "start";
    div.style["justify-content"] = "end";
    div.setAttribute("style", "place-content: initial");

    test(function() {
        assert_equals(div.style["align-content"],
            "initial", "place-content specified value for align-content");
    }, "Check place-content: initial - align-content expanded value");

    test(function() {
        assert_equals(div.style["justify-content"],
            "initial", "place-content specified value for justify-content");
    }, "Check place-content: initial - justify-content expanded value");
</script>
</body></html>
